bio_knowledge_miner/requirements.txt
requests
beautifulsoup4
semantic-scholar-api
pubmed-lookup
PyMuPDF
pdfplumber
paddlepaddle-gpu
paddleocr
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
openai
anthropic
google-generativeai
python-dotenv
neo4j
tqdm
crossrefapi